pyproject.toml
[project]
name = "iot-server"
version = "0.1.0"
description = "Iot Server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "celery>=5.5.3",
    "fastapi[standard]>=0.121.2",
    "jinja2>=3.1.6",
    "mysqlclient>=2.2.7",
    "python-dotenv>=1.0.0",
    "redis>=7.0.1",
    "sqlmodel>=0.0.30",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.10",
]
